Make some RTS-GMLC fields optional or more flexible.
Shunt-related columns can be omitted if they are never used.

There is a variable number of fuel-related columns.  The files in the RTS-GMLC repository
use 5 columns, but the last of these 5 columns is never used (it is always 'NA').  The
parser was hard-coded to read 4 columns and ignore the 5th.  While this works for the
repository files, it is overly rigid.  The new code allows any number of columns, using
whatever columns are present and appropriately populated for each generator.