When parsing RTS-GMLC files, filter which scalar reserve requirements are included in a model based on the simulation type and what's listed in the Reserve_Products row of simulation_objects.csv.
Note that this only applies to scalar reserve requirements (the ones listed in
reserves.csv). It does not filter timeseries by reserve type; if a timeseries is listed
in timeseries_pointers.csv for a given reserve, the timeseries is included in the model
even if the reserve type isn't listed in simulation_objects.csv.

If simulation_objects.csv does not have a Reserve_Products row then all reserve products
are accepted for both DAY_AHEAD and REAL_TIME models.