Add flag to ignore look-ahead period counts when parsing RTS-GMLC.
RTS-GMLC metadata has "Look_Ahead_Periods_per_Step" fields that indicate
how much data to include in each simulation step. Consumers of RTS-GMLC
data (such as Prescient) may have their own mechanisms for determining
the simulation horizon, and it may not match the values stored in the
metadata.

This commit adds a new argument to rts_gmlc.parser.parse_to_cache(), called
"honor_lookahead". If this is True (the default), it uses the lookahead
period count found in the data's metadata. If it is False, it does not
include any extra data beyond the specified end_time. When calling this
function, consumers of the parsed data are expected to adjust the end_time
to include any look-ahead periods it may need.